Rename the file names
