+++ tests/unit/CMakeLists.txt
@@ -48,6 +48,7 @@ set(TEST_SOURCES
     test_event_binner.cpp
     test_filter.cpp
     test_fvm_multi.cpp
+    test_ips_cell.cpp
     test_mc_cell_group.cpp
     test_lexcmp.cpp
     test_mask_stream.cpp
@@ -60,8 +61,7 @@ set(TEST_SOURCES
     test_optional.cpp
     test_mechinfo.cpp
     test_partition.cpp
-    test_path.cpp
-    test_pi_cell.cpp
+    test_path.cpp   
     test_point.cpp
     test_prefixbuf.cpp
     test_probe.cpp
